Fixed query for getting spaces #143

+++ src/main/java/com/klai/stl/service/statistics/dto/SpaceDTO.java
@@ -1,5 +1,6 @@
 package com.klai.stl.service.statistics.dto;
 
+import java.math.BigInteger;
 import javax.persistence.Tuple;
 import lombok.Value;
 
@@ -14,7 +15,7 @@ public class SpaceDTO {
     public SpaceDTO(Tuple tuple) {
         this.name = tuple.get("name", String.class);
         this.description = tuple.get("description", String.class);
-        this.photos = tuple.get("photos", Integer.class);
-        this.coordinates = tuple.get("coordinates", Integer.class);
+        this.photos = tuple.get("photos", BigInteger.class).intValue();
+        this.coordinates = tuple.get("coordinates", BigInteger.class).intValue();
     }
 }
